Refactor unittest
