+++ tests/test_models/test_roi_heads/__init__.py
@@ -1,0 +1,3 @@
+from .utils import _dummy_bbox_sampling
+
+__all__ = ['_dummy_bbox_sampling']
