+++ tests/test_utils/test_misc.py
